requirements.txt
uvicorn~=0.17.6
fastapi~=0.75.0
mysql-connector-python~=8.0.29
python-dotenv~=0.20.0
boto3~=1.24.12
PyJWT~=2.8.0